Update doc_creator.py
createDoc and getFile Now work correctly. This includes separating the get and create commands to the server they were called in. Meaning if a file is created in server A and you try to get that file in server B, the file will not be visible unless you call getFile in server A