Update Reminder.py
Reminder now works as a slash command but because of its implementation it is easier for the user to use but it also has less control of the time. To deal with this there is now complex reminder as well. It takes more arguments but allows for fine tuning of the time. This gives the user a choice, a quicker timer setup if the exact time is not required. Or a longer timer setup if they want something specific